Figured out why PID Controller wasn't working
Because we were using FieldCentric, the values of the controller were in relation to the field. However, the chassis_speeds are with respect to the robot themselves; two different reference points. For this reason, without a complicated mathematical conversion between the two coordinate systems, you cannot use the PID directly here. I looked for a function that could do the translation directly, but was unable to find. If we need to use the PID, we'll have to develop the invert function for the field velocity